requirements.txt
boto3==1.28.53
pillow==10.0.1
rembg==2.0.50
starlette==0.31.1
uvicorn[standard]==0.23.2
